9_vector_databases/5_chromadb/requirements.txt
chromadb
langchain
langchain-openai
langchain_chroma
python-dotenv